requirements.txt
Django==1.7
Markdown==2.3
Minidetector==0.1
Pillow==2.5.3
Whoosh==2.6.0
django-appconf==0.6
django-common-helpers==0.6.2
django-compressor==1.4
django-crispy-forms==1.4.0
django-cron==0.3.4
django-debug-toolbar==1.2.1
django-djconfig==0.1.5
django-geojson==2.6.0
django-haystack==2.2.1
django-infinite-scroll-pagination==0.1.1
django-leaflet==0.14.1
django-lockdown==1.0
mod-wsgi==4.3.0
mod-wsgi-metrics==1.1.0
psutil==2.1.1
psycopg2==2.5.4
pytz==2014.7
six==1.8.0
sqlparse==0.1.11
wsgiref==0.1.2
